Update to User object / Server side saving initial implementation
Users can now have friends.  Methods for adding friends, encountered
humons, and new humon instances are also added.  Additionally a isDirty
field was added.  This field is set to true anytime data about the
user changes.  Using the isDirty will allow us to know when to write
back to the database and when to not.

Basic save logic was added.  After 5 minutes of interaction with the
server, it will check to see if data is dirty and write it back to the
database if it is.  This save logic is blocked by waiting to read from
the user (so it may be longer than 5 minutes before a save) but saves
will always occur os connection close.

A Database class was also created.  Its purpose is to check the setup
of the database before the server will allow any connections.  This
class also allows for the tables to be cleared, as well as for test
data to be populated <future implementations.>

Signed-off-by: Tom Schutte <tjschutte@wisc.edu>