Saving of server response to file
When a user registers or logs in, it is assumed that the server always has the
correct copy of data.  This data is written (or overwrites) the current copy
of the users data.  From login on the client is assumed to have the correct
copy of data which will be written back to the server later.

Signed-off-by: Tom Schutte <tjschutte@wisc.edu>