User object is now saved back to server with the party
Updated and cleaned up save service.  Previously is was a
'cart before the horse' situation where humons were the only thing
it was meant to save.  Have since fixed so it is a bit more generic
and can handle saving the user back to the server as well.

TODO: Update the user object so we are sending an updated file to
the server!

Signed-off-by: Tom Schutte <tjschutte@wisc.edu>