GetHumon and GetInstance implementations.
GetHumon now will return a humon object.  This command can take either a
hID or iID and will return the corrent humon object regardless.

GetInstance now will return a humon object with the relevant instance data
filled in.

Both of these commands are part of the implemetation of network battles.

Signed-off-by: Tom Schutte <tjschutte@wisc.edu>