Battle Accepted Command
Needed a way for when a user accepts a battle to send a notification
to the one who issued the battle to trigger the battle state in both
devices.  This makes battles a sort of 3 way hand-shake ordeal so that
we end up putting both users into the battle state at an appropriate time.

Signed-off-by: Tom Schutte <tjschutte@wisc.edu>