Online User History tracking
Created class to keep track of users who are currently marked as online.
This will be used as a means of dissociating the server from the client to
a more connectionless service.

Old flow:
User login and connection is maintained.  Not ideal for client as it
requires an always on connection which on mobile devices are not always
stable.

New flow:
User logs in, and disconnects.
On further actions the user includes who they are with each action.  This
is verified against a list of logged in users the server will keep track
of.  This also means that at the end of a session a logout command needs
to be issued, or the server will need to have a limit on how long to wait
before logging a user out.

Signed-off-by: Tom Schutte <tjschutte@wisc.edu>